backend/requirements.txt
Django==1.10.3
psycopg2==2.6.2
djangorestframework==3.5.3
django-pdb==0.5.1
